pyproject.toml
[project]
name = "cosmos-chessbot"
version = "0.1.0"
description = "Cosmos-Reason2 + Stockfish + PPO chess manipulation system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "av>=16.1.0",
    "pillow>=12.0.0",
    "transformers>=4.57.3",
    "torch>=2.9.0",
    "torchvision",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.0.0",
    "onnxruntime>=1.16.0",
    "ultralytics>=8.0.0",
    "python-chess>=1.10.0",
    "python-dotenv>=1.0.0",
    "peft>=0.7.0",
    "datasets>=2.14.0",
    "bitsandbytes>=0.41.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "lerobot>=0.3.2",
    "mcp[cli]>=1.2.0",
    "scipy>=1.10.0",
]

[project.scripts]
cosmos-chessbot = "cosmos_chessbot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
